+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ## [Unreleased]
 ### Changed
 - Update username on leaderboard and tournament records when processing a score update.
+- Automatically stop empty authoritative matches after a configurable amount of time.
 
 ## [2.9.1] - 2020-01-14
 ### Changed